requirements.txt
astroid==2.2.5
autopep8==1.5
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.4
django-bootstrap-modal-forms==1.4.4
django-crispy-forms==1.7.2
django-filter==2.2.0
django-paypal==1.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
get==2019.4.13
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
paypal==1.2.5
pbr==5.4.4
Pillow==6.2.0
post==2019.4.13
psycopg2==2.8.4
public==2019.4.13
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
